refactor: team and member logic and ui to components
